requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
Django==3.0.6
inflect==4.1.0
mysqlclient==1.4.6
pytz==2020.1
sqlparse==0.3.1
djangorestframework==3.10.3
